% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaksim.R
\name{outbreaksim}
\alias{outbreaksim}
\title{outbreaksim

\code{outbreak simulation functions}}
\usage{
outbreaksim(n_sample = 20, ita = log(0.01), theta = log(0.01), z_m = 0,
  z_sig = 0, as = 10, ms = 1, fixed_beta = 1, plotsir = T, sim = 2,
  seq_length = 1000, transi_m = NULL, transv_m = NULL)
}
\arguments{
\item{n_sample}{sample size}

\item{ita}{ita, fixed value, \code{beta = exp(ita + theta * z)}}

\item{theta}{theta, fixed value, \code{beta = exp(ita + theta * z)}}

\item{z_m}{mean for z, \code{beta = exp(ita + theta * z)}}

\item{z_sig}{deviation for z, \code{beta = exp(ita + theta * z)}}

\item{as}{shape for sampling time interval gamma distribution}

\item{ms}{mean for sampling time interval gamma distribution}

\item{fixed_beta}{when \code{sim=1}, beta is set as a fixed value}

\item{plotsir}{plot the ground truth of SIR model when set as True}

\item{sim}{simulation pattern, when \code{sim=1}, the rate of exponential distribution (dist. of transmission interval) beta is set as constant, when \code{sim=2}, assume log(beta) follows a normal distribution}

\item{seq_length}{length of simulated sequence}

\item{transi_m}{rate of transition mutation, if NULL, it is set as 0.1/seq_length}

\item{transv_m}{rate of transvertion mutation, if NULL, it is set as 1/transv_m}
}
\description{
outbreaksim is a tool to simulate an outbreak based on simple assupmtions
on the transmission, sampling and mutation
}
\author{
yixuan ye
}
