% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqsim.R
\name{seqsim}
\alias{seqsim}
\title{seqsim}
\usage{
seqsim(seq_length, transi_m, transv_m, n_sample, t, h)
}
\arguments{
\item{seq_length}{length of simulated sequence}

\item{transi_m}{rate of transition mutation}

\item{transv_m}{rate of transvertion mutation}

\item{n_sample}{sample size}

\item{t}{true infection time}

\item{h}{true infector}
}
\description{
function to simulate sequence during an outbreak
}
