% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreaksim.R
\name{outbreaksim}
\alias{outbreaksim}
\title{outbreaksim

\code{outbreak simulation functions}}
\usage{
outbreaksim(n_sample = 20, ita = 0.1, theta = 0.1, z_m = 0, z_sig = 1,
  as = 3, ms = 1, fixed_beta = 1, plotsir = T, sim = 2,
  seq_length = 10000, transi_m = NULL, transv_m = NULL)
}
\arguments{
\item{n_sample}{sample size}

\item{ita}{ita}

\item{theta}{theta}

\item{z_m}{mean for z}

\item{z_sig}{deviation for z}

\item{as}{shape for sampling time interval distribution}

\item{ms}{mean for sampling time interval distribution}

\item{fixed_beta}{when sim=1, the value for beta}

\item{plotsir}{plot the ground truth of SIR model or not}

\item{sim}{when sim=1, the beta is set as constant, when sim=2, beta follows a distribution}

\item{seq_length}{length of simulated sequence}

\item{transi_m}{rate of transition mutation}

\item{transv_m}{rate of transvertion mutation}
}
\description{
outbreaksim is a tool to simulate an outbreak based on simple assupmtions
on the transmission, sampling and mutation
}
\author{
yixuan ye
}
